Non-tech edit complete
@sammyd @jawwad to the next!
I cleared out all the TODOs I could and left one for each of you.

@sammyd - where’s the best place for feedback?  I actually don’t have
too much because it was really clean and easy to work with. I did do
some fairly intensive rewriting, but it was typical stuff like reducing
redundancy, livening up the language and edits for brevity. Happy to go
more in depth if @jawwad would like more detail, but it’s probably more
efficient to finish chapter 7 and roll it up at once.