Non-tech edit done. Really
Okay, sorry about the multiple commits. I realized that I had worked on
two versions of the copy while editing and the previous version
uploaded did not include a good chunk of my work. I combined the files,
replaced the code with what was here before I started and reviewed
again. Should be good to go.