Remove the teaser image of the app in favor of showing the user just the screenshot with the alignment issue (with updated styling)
Removed it for a few reasons. For one the user isn't implementing the rating view. Also the text says that its the finished version of the app but it already looks like that now. Also the new image will show the first view with the issue so the other partial screenshot that was showing the issue can be removed.